refactor: 의존관계 리펙토링
