+++ tomcat/src/main/java/nextstep/jwp/SessionManager.java
@@ -1,12 +1,13 @@
-package org.apache.coyote.http11;
+package nextstep.jwp;
 
 import org.apache.catalina.Manager;
-import java.util.HashMap;
+import org.apache.coyote.http11.Session;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class SessionManager implements Manager {
 
-    private static final Map<String, Session> SESSIONS = new HashMap<>();
+    private static final Map<String, Session> SESSIONS = new ConcurrentHashMap<>();
 
     @Override
     public void add(final Session session) {
